Updated Backend - Shapes

+++ Back End/Paint-BE/src/main/java/ShapesPackage/Hexagon.java
@@ -2,7 +2,7 @@ package ShapesPackage;
 
 public class Hexagon extends Shape{
 	
-	private double sides;
+	private int sides;
 	private double radius;
 	
 	public Hexagon() {
@@ -26,11 +26,11 @@ public class Hexagon extends Shape{
 		return new Hexagon(this);
 	}
 
-	public double getSides() {
+	public int getSides() {
 		return sides;
 	}
 
-	public void setSides(double sides) {
+	public void setSides(int sides) {
 		this.sides = sides;
 	}
